perf(doctor): split criterion summary and baseline tooling tests
