+++ .github/workflows/criterion-bench.yml
@@ -13,6 +13,7 @@ on:
       - "bench/criterion-ab.mjs"
       - "bench/criterion-baselines.mjs"
       - "bench/criterion-impact.mjs"
+      - "bench/criterion-summary.mjs"
       - "bench/dialect-guard.mjs"
       - "bench/generate.mjs"
       - "rust-toolchain.toml"
